refactor: Break down message_handlers.py into domain-specific handler files
